refactor: refactor code and remove usings
