+++ src/Serendipity.Domain/Constants/States.cs
@@ -1,0 +1,15 @@
+﻿using System.Runtime.Serialization;
+
+namespace Serendipity.Domain.Constants;
+
+public enum States
+{
+    [EnumMember(Value = nameof(Running))]
+    Running,
+    [EnumMember(Value = nameof(Sleeping))]
+    Sleeping,
+    [EnumMember(Value = nameof(Walking))]
+    Walking,
+    [EnumMember(Value = "Sitting")]
+    Sitting
+}
