+++ src/Serendipity.Domain/Interfaces/Repository/IReportRepository.cs
@@ -1,9 +1,4 @@
 ﻿using Serendipity.Domain.Contracts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Serendipity.Domain.Models;
 
 namespace Serendipity.Domain.Interfaces.Repository;
