+++ src/Serendipity.Domain/Models/Device.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Serendipity.Domain.Models;
 
 public class Device
