+++ src/Serendipity.Domain/Services/ReportService.cs
@@ -1,9 +1,4 @@
-﻿using System.Net;
-using Amazon;
-using Amazon.S3;
-using Amazon.S3.Model;
-using Microsoft.Extensions.Configuration;
-using Serendipity.Domain.Contracts;
+﻿using Serendipity.Domain.Contracts;
 using Serendipity.Domain.Interfaces.Repository;
 using Serendipity.Domain.Interfaces.Services;
 using Serendipity.Domain.Models;
