+++ src/Serendipity.Infrastructure/Models/User.cs
@@ -1,5 +1,4 @@
 using Microsoft.AspNetCore.Identity;
-using Microsoft.VisualBasic;
 
 namespace Serendipity.Infrastructure.Models;
