+++ src/Serendipity.WebApi/Controllers/DeviceController.cs
@@ -4,7 +4,6 @@ using Microsoft.AspNetCore.Mvc;
 using Serendipity.Domain.Contracts;
 using Serendipity.Domain.Interfaces.Services;
 using Serendipity.Infrastructure.Models;
-using Serendipity.WebApi.Contracts;
 using Serendipity.WebApi.Contracts.Requests;
 using Serendipity.WebApi.Contracts.Responses;
 using Serendipity.WebApi.Filters;
