+++ src/Serendipity.WebApi/Extensions/WebApplicationExtensions.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Identity;
 using Serendipity.Domain.Defaults;
-using Serendipity.Infrastructure.Database;
 using Serendipity.Infrastructure.Models;
 
 namespace Serendipity.WebApi.Extensions;
